Handle users with no portal authentiaction.[#97849058]

+++ app/controllers/import_controller.rb
@@ -48,25 +48,29 @@ class ImportController < ApplicationController
 
     portal = Concord::AuthPortal.portal_for_url(request_json[:portal_url]) 
     authentication = Authentication.find_by_provider_and_uid(portal.strategy_name,request_json[:domain_uid])
-    user = User.find(authentication.user_id)
+    if authentication
+      user = User.find(authentication.user_id)
 
-    import_activity = LightweightActivity.import(request_json[:activity],user)
+      import_activity = LightweightActivity.import(request_json[:activity],user)
 
-    req_url = "#{request.protocol}#{request.host_with_port}"
-    response_add_to_portal = import_activity.portal_publish(user,portal,req_url)
-    response_code = response_add_to_portal.code
-    if response_code == 201
-      response_publish = import_activity.republish_for_portal(portal,req_url)
-      response_code = response_publish.code
+      req_url = "#{request.protocol}#{request.host_with_port}"
+      response_add_to_portal = import_activity.portal_publish(user,portal,req_url)
+      response_code = response_add_to_portal.code
+      if response_code == 201
+        response_publish = import_activity.republish_for_portal(portal,req_url)
+        response_code = response_publish.code
+      else
+        import_activity.destroy
+        response.headers["data"] = {:response_code => response_code}.to_json
+        render :nothing => true and return
+      end
+
+      response_body = JSON.parse "#{response_publish.body}", :symbolize_names => true
+      response.headers["data"] = {:external_activity_id => response_body[:activity_id],
+                                  :response_code => response_code}.to_json
     else
-      import_activity.destroy
-      response.headers["data"] = {:response_code => response_code}.to_json
-      render :nothing => true and return
+      response.headers["data"] = {:response_code => 401}.to_json
     end
-
-    response_body = JSON.parse "#{response_publish.body}", :symbolize_names => true
-    response.headers["data"] = {:external_activity_id => response_body[:activity_id],
-                                :response_code => response_code}.to_json
     render :nothing => true
   end
 end
